+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.5.0
 * `NEW` setting `runtime.unicodeName`
+* `NEW` fully supports `---@generic T`
 * `FIX` [#274](https://github.com/sumneko/lua-language-server/issues/274)
 
 ## 1.4.0